+++ packages/jest-runner/src/types.ts
@@ -43,13 +43,13 @@ export type TestRunnerOptions = {
 
 // make sure all props here are present in the type below it as well
 export type TestRunnerContext = {
-  changedFiles?: Set<Config.Path>;
-  sourcesRelatedToTestsInChangedFiles?: Set<Config.Path>;
+  changedFiles?: Set<string>;
+  sourcesRelatedToTestsInChangedFiles?: Set<string>;
 };
 
 export type TestRunnerSerializedContext = {
-  changedFiles?: Array<Config.Path>;
-  sourcesRelatedToTestsInChangedFiles?: Array<Config.Path>;
+  changedFiles?: Array<string>;
+  sourcesRelatedToTestsInChangedFiles?: Array<string>;
 };
 
 // TODO: Should live in `@jest/core` or `jest-watcher`